+++ examples/discovery-components-react-example/src/App.js
@@ -2,8 +2,12 @@ import React from 'react';
 import DiscoveryV1 from 'ibm-watson/discovery/v1';
 import './app.scss';
 import refinementsQueryResponse from './fixtures/refinementsQueryResponse';
-
-import { DiscoverySearch, SearchInput, SearchRefinements } from '@disco-widgets/react-components';
+import {
+  DiscoverySearch,
+  SearchInput,
+  SearchRefinements,
+  ResultsPagination
+} from '@disco-widgets/react-components';
 
 const App = () => {
   // TODO: this is a dummy client to route requests to the server since CP4D doesn't support CORS
@@ -25,6 +29,7 @@ const App = () => {
         labelText={'This is some label text...'}
       />
       <SearchRefinements queryResponse={refinementsQueryResponse} />
+      <ResultsPagination />
     </DiscoverySearch>
   );
 };