+++ examples/discovery-components-react-example/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.1](https://github.ibm.com/Watson-Discovery/disco-widgets/compare/discovery-components-react-example@1.11.0...discovery-components-react-example@1.11.1) (2019-09-20)
+
+**Note:** Version bump only for package discovery-components-react-example
+
 # [1.11.0](https://github.ibm.com/Watson-Discovery/disco-widgets/compare/discovery-components-react-example@1.10.0...discovery-components-react-example@1.11.0) (2019-09-19)
 
 ### Features