+++ packages/discovery-components-react/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.16.0](http:///discovery-components-react/compare/@disco-widgets/react-components@1.15.0...@disco-widgets/react-components@1.16.0) (2019-09-20)
+
+### Features
+
+- export search context ([d9bed14](http:///discovery-components-react/commits/d9bed14))
+
 # [1.15.0](http:///discovery-components-react/compare/@disco-widgets/react-components@1.14.0...@disco-widgets/react-components@1.15.0) (2019-09-19)
 
 ### Features